refactor(analysis): extract qualify stage to stages/qualify.py
Moves generate_qualitative and its 6 supporting functions verbatim into
stages/qualify.py + stages/_qualify_fallbacks.py (split to satisfy the
300-line new-file cap: qualify.py=222 lines, _qualify_fallbacks.py=227 lines).

Re-exports all 7 names from deep_analysis_pipeline for backward compat.
Updates 5 test patch targets to reflect the new module locations.
deep_analysis_pipeline.py shrinks from 902 to 507 lines.
