+++ tests/unit/analysis/test_deep_analysis_pipeline.py
@@ -302,7 +302,7 @@ class TestGenerateQualitative:
         mock_crew_instance.kickoff.return_value = mock_crew_result
 
         mocker.patch(
-            "finwiz.analysis.deep_analysis_pipeline._get_analysis_crew",
+            "finwiz.analysis.stages.qualify._get_analysis_crew",
             return_value=mock_crew_instance,
         )
 
@@ -328,11 +328,11 @@ class TestGenerateQualitative:
         mock_crew_instance.kickoff.return_value = mock_crew_result
 
         mocker.patch(
-            "finwiz.analysis.deep_analysis_pipeline._get_analysis_crew",
+            "finwiz.analysis.stages.qualify._get_analysis_crew",
             return_value=mock_crew_instance,
         )
         mocker.patch(
-            "finwiz.analysis.deep_analysis_pipeline._extract_qualitative",
+            "finwiz.analysis.stages.qualify._extract_qualitative",
             return_value=mocker.MagicMock(),
         )
 
@@ -457,11 +457,11 @@ class TestAnalyzeHolding:
             return_value=(mock_deep_analysis_result, mock_quant),
         )
         mocker.patch(
-            "finwiz.analysis.deep_analysis_pipeline.generate_qualitative",
+            "finwiz.analysis.stages.qualify.generate_qualitative",
             return_value=mock_qualitative_insights,
         )
         mocker.patch(
-            "finwiz.analysis.deep_analysis_pipeline._safe_strategic",
+            "finwiz.analysis.stages.qualify._safe_strategic",
             return_value=None,
         )
 
@@ -521,11 +521,11 @@ class TestAnalyzeHolding:
             return_value=(mock_result, mock_quant),
         )
         mocker.patch(
-            "finwiz.analysis.deep_analysis_pipeline.generate_qualitative",
+            "finwiz.analysis.stages.qualify.generate_qualitative",
             return_value=mock_qualitative_insights,
         )
         mocker.patch(
-            "finwiz.analysis.deep_analysis_pipeline._safe_strategic",
+            "finwiz.analysis.stages.qualify._safe_strategic",
             return_value=None,
         )
